Settings

+++ src/app/tasks/order-list/order-list.component.ts
@@ -1,8 +1,10 @@
 import {Component, OnInit} from '@angular/core';
 import {AngularFirestore} from '@angular/fire/firestore';
+import {MatDialog} from '@angular/material';
 import {Router} from '@angular/router';
 import {FirebaseAuthService} from '../../auth/firebase-auth.service';
-import {NotificationService} from '../../notification.service';
+import {ConfirmDialogComponent} from '../../common/confirm-dialog/confirm-dialog.component';
+import {NotificationService} from '../../common/notification.service';
 import {Order} from '../order.model';
 
 @Component({
@@ -18,14 +20,15 @@ export class OrderListComponent implements OnInit {
     constructor(private router: Router,
                 private db: AngularFirestore,
                 private auth: FirebaseAuthService,
-                private notificationService: NotificationService) {
+                private notificationService: NotificationService,
+                private dialog: MatDialog) {
     }
 
     ngOnInit() {
         this.auth.user$.subscribe(user => {
             this.teamId = user.teamId;
             this.db.collection('orders', ref => ref.where('teamId', '==', this.teamId)).get().subscribe(
-                orders => orders.forEach(o => this.orders.push( {...o.data() as Order, id: o.id}))
+                orders => orders.forEach(o => this.orders.push({...o.data() as Order, id: o.id}))
             );
         });
     }
@@ -35,11 +38,16 @@ export class OrderListComponent implements OnInit {
     }
 
     delete(orderId: string) {
-        this.db.collection('/orders').doc(orderId).delete().then(() => {
-            this.notificationService.deletedSuccessfully();
-            this.orders = this.orders.filter(o => o.id !== orderId);
-        })
-            .catch(() => this.notificationService.error());
+        const dialogRef = this.dialog.open(ConfirmDialogComponent, {width: '25%'});
+        dialogRef.afterClosed().subscribe(result => {
+            if (result) {
+                this.db.collection('/orders').doc(orderId).delete().then(() => {
+                    this.notificationService.deletedSuccessfully();
+                    this.orders = this.orders.filter(o => o.id !== orderId);
+                })
+                    .catch(() => this.notificationService.error());
+            }
+        });
     }
 
     edit(order: Order) {
